Добавил репозиторий с городами и вывод их на экран

+++ app/src/main/java/com/skysoft/skyweather/model/City.kt
@@ -1,3 +1,3 @@
 package com.skysoft.skyweather.model
 
-data class City (val name: String, val latitude: Float, val longitude: Float)
+data class City(val name: String, val latitude: Double, val longitude: Double)
